requirements.txt
streamlit
python-dotenv
openai
streamlit-mermaid